pyproject.toml
[build-system]
requires = [ "poetry-core",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "moatless"
version = "0.0.3"
description = ""
authors = [ "Albert Örwall <albert@moatless.ai>",]
readme = "README.md"

[tool.ruff]
target-version = "py310"
extend-exclude = [ "tests", "evaluations", "notebooks",]

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
pydantic = "2.8.2"
datasets = "2.20.0"
litellm = "^1.44.22"
instructor = "^1.3.7"
tiktoken = "^0.7.0"
networkx = "3.3"
tree-sitter = "0.22.3"
tree-sitter-python = "0.21.0"
tree-sitter-java = "0.21.0"
rapidfuzz = "3.9.5"
openai = "^1.41.0"
anthropic = "^0.34.1"
llama-index = "0.10.65"
llama-index-embeddings-openai = "0.1.11"
llama-index-embeddings-voyageai = "0.1.4"
llama-index-readers-file = "0.1.33"
jsonref = "1.1.0"
pylint = "3.2.6"
numpy = ">=1.0,<2.0"
scipy = "1.14.0"
faiss-cpu = "1.8.0.post1"
gitpython = "3.1.43"
pyarrow = "17.0.0"
requests = "2.32.3"
pandas = "2.2.2"
plotly = "^5.24.0"
pygraphviz = "^1.13"
graphviz = "^0.20.3"
tabulate = "^0.9.0"
streamlit = "^1.38.0"
matplotlib = "^3.9.2"

[tool.ruff.lint]
select = [ "B", "DTZ", "E", "F", "I", "LOG", "N", "PLE", "SIM", "T20", "UP",]
ignore = [ "E501", "F401", "UP007" ]

[tool.poetry.group.dev.dependencies]
pytest = "8.3.2"
pytest-mock = "3.14.0"
python-dotenv = "1.0.1"
mypy = "1.10.0"
ruff = "0.5.5"
coverage = "^7.6.1"
